+++ src/library/payfine/PayFineUI.java
@@ -7,12 +7,12 @@ public class PayFineUI {
 
 	public static enum uI_sTaTe { INITIALISED, READY, PAYING, COMPLETED, CANCELLED };
 
-	private pAY_fINE_cONTROL CoNtRoL;
+	private PayFineControl CoNtRoL;
 	private Scanner input;
 	private uI_sTaTe StAtE;
 
 	
-	public PayFineUI(pAY_fINE_cONTROL control) {
+	public PayFineUI(PayFineControl control) {
 		this.CoNtRoL = control;
 		input = new Scanner(System.in);
 		StAtE = uI_sTaTe.INITIALISED;
